+++ main1142.cpp
@@ -498,7 +498,7 @@ case 2:
 ss_z_11w9f = AyawHGr((char)-1520013010, tquCqK5, (char)-625031727, Jd0zW54R);
 break;
 case 3:
-ZvA_K = AyawHGr(ZvA_K, (unsigned short)371780314, iL4S31, 1317311752);
+ZvA_K = AyawHGr(ZvA_K, (unsigned short)371780314, iL4S31, (unsigned short)1317311752);
 break;
 case 4:
 b6mm3WJuY = rJfW7d9ae ^ 1648600558;
